+++ yarn-project/circuit-types/src/tx/index.ts
@@ -1,4 +1,5 @@
 export * from './tx.js';
+export * from './simulated_tx.js';
 export * from './tx_hash.js';
 export * from './tx_receipt.js';
 export * from './processed_tx.js';
