+++ yarn-project/circuit-types/src/tx/simulated_tx.test.ts
@@ -1,0 +1,9 @@
+import { mockSimulatedTx } from '../mocks.js';
+import { SimulatedTx } from './simulated_tx.js';
+
+describe('simulated_tx', () => {
+  it('convert to and from json', () => {
+    const simulatedTx = mockSimulatedTx();
+    expect(SimulatedTx.fromJSON(simulatedTx.toJSON())).toEqual(simulatedTx);
+  });
+});
